LayoutTests/imported/w3c/web-platform-tests/workers/shared-worker-in-data-url-context.window-expected.txt

CONSOLE MESSAGE: Cannot load http://localhost:8800/workers/support/post-message-on-load-worker.js.

PASS Create a shared worker in a data url frame
PASS Create a data url shared worker in a data url frame
